Integrate us_web_design_standards
This is just the first step. The dependency is in the gemspec, the generator
and sass modules depend on us_web_design_standards to integrate that gem's
generator and sass modules, and the guides_style_18f.scss file imports the
us_web_design_standards styles.

However, this commit does not update the layouts and include templates to use
the us_web_design_standards styles.