Fix child nav broken in #28, sidebar nesting
Forgot that the sidebar depended on the `parent:` property. This fixes the
problem by introducing the `guides_style_18f_should_expand_nav` tag that
allows the template to compare this value against the parent URL.

Also implements arbitrarily deep sidebar nesting. Requires a couple of
plugins, given the limitations of the basic Liquid template syntax. A style
update may perhaps be in order.