Breadcrumb implementation
Uses the `navigation:` data (generated by `./go update_nav` and stored in
`_config.yml`) to produce a mapping from page URL to a list of breadcrumb
objects. `Generator.generate` attaches this data to each page, and the
`includes/breadcrumbs.html` template renders it as an unordered list.

The breadcrumbs are not yet included and styled in the standard layout, but
can be included using the following Liquid directive:

  {% guides_style_18f_include breadcrumbs.html %}