Add git email and name config
This was causing
ClearTemplateFilesAndCreateNewRepositoryTest.create_initial_repo
to fail on Travis. This was especially difficult to debug because the
SystemExit caused the rake test process to exit immediately without generating
any output. The SystemExit rescue statement introduced here rectifies this,
including the log file output in a proper Minitest assertion failure.